Stored Alarms
In this fix the alarm system is able to track all alarms sent out so that in the future you can see the alarm history of a given structure. In addition to this major feature this version also includes a big optimization fix, tremendously speeding up structure loading time and data retrieval. A small feature was also implemented and that was the inclusion of a link back to the old SOS Dashboard.